pyproject.toml
[project]
name = "ai-shark"
version = "0.1.0"
description = "VC Document Analyzer - AI-powered pitch deck analysis tool"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.28.0",
    "google-generativeai>=0.3.0",
    "PyMuPDF>=1.23.0",
    "python-pptx>=0.6.21",
    "Pillow>=10.0.0",
    "python-dotenv>=1.0.0",
]
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[project.scripts]
vc-analyzer = "src.main:main"